refactor: move loading tests to separate files
